warehouse_scada_security/ida-agent/requirements.txt
requests
cryptography
pyyaml
python-dotenv
boto3